Password field added to user registration -i18
config.php:
-Added encryption_salt value

User controller,pending_users model, and releated view files:
Added password field to the form along withsome javascript to make sure
the field is filled out correctly. When the user submits the form to the
user controller the password is hashed using hash() function of php, as
well as a salt value in config.php ($config['encryption_salt']). The
hash value is stored into the database.

Readme:
-Updated README to explain encryption_salt value