Updated javascript form validation. Tags are formmatted by commas on client side.
global.js changes:
-Updated word_count function to take a optional delimiter character.
-Added format_list[seperation_string][replace_string] rule. This
replaces all instances of seperation_string with replacement_string.
If more then 1 replace_string is next to eachother then the group is
replaced to just 1 instance of replacement_string. Also, if the
replacement_string is at the beginning or end of the input, it is
removed.

views/topic/Submit.js changes:
-Added code to delimit tag field by commas on client side, but send
field as delimited by a space.

views/topic/Submit.php changes:
-Added rstatement that tags must be at least 4 characters long.